开放文件广播文件删除接口、文件排序接口、文件列表重置接口，实现文件广播页面文件排序功能，音频导入与歌单导入暂未实现

+++ audioweb-common/src/main/java/com/audioweb/common/config/Global.java
@@ -40,7 +40,7 @@ public class Global
     private static boolean addressEnabled;
     
     /** 删除或丢失音频文件保存记录天数 */
-    private static Integer lostFile;
+    private static Long lostFile;
    
     public static String getName()
     {
@@ -159,11 +159,11 @@ public class Global
 		Global.wordPath = wordPath;
 	}
 
-	public static Integer getLostFile() {
+	public static Long getLostFile() {
 		return lostFile;
 	}
 
-	public void setLostFile(Integer lostFile) {
+	public void setLostFile(Long lostFile) {
 		Global.lostFile = lostFile;
 	}
 }
